Initial documentation commit
This initial commit contains a starting point for an "AFM/SPM
Components" document, where I diagram out the different components
that exist in AFM/SPM systems. The main purposes are:
(a) To understand all of this (it is quite confusing);
(b) To list/comprehend all of the different parameters (which will help
when standardizing them).

Note:
- The .org file contains all the code (primarily plantuml documents). It
tangles this code into the code subdirectry, which allows us to re-use
code in different code blocks.
- The images subdirectory contains diagrams created from this code.