[style] modify plantuml to consistent format
Particularly, we removed (as much as possible) explicit direction/
location logic. I realized (through trial and error), what appears
to be the ideal mechanism to set up structure in plantuml
1. Design your structure to be going top-to-bottom, as this appears
to be the default? So, in most situations, you will want ot be using
--> instead of ->, to indicate up/down arrows.
2. Use 'package " " <<layout>>' to bundle/group components. This is
particularly useful if you are dealing with stuff that should be
organized in a particular direction. Note that our style.puml
sets packages of stereotype 'layout' to have hidden borders.
3. Use left/right arrows sparingly, where it makes sense to guide
graphviz.
4. Alternate the direction your graph is using by adding 'left to right
direction' at the top, if needed. This gives you the option of which
structure you prefer. It's better to design it top/down and switch to
this, because otherwise the arrows logic gets flipped and your brain
gets confused.