feat: major update to afm-spm components doc
- A lot of work was done to standardize the different puml diagrams
and allow them to be interfaced reasonable. For example, for all
'simple' diagrams (ones hooked together), all input/output names
have had a component prefix and in/out post-fix, for clarity.
- Adding a section for example AFM configurations: contact mode,
am-afm, and fm-afm.
- Adding section where we list all of the potential metadata parameters.
- For style.puml, we switched from skinparam to (mostly) the new style
approach.
- For the html export, we returned to toc (with a limit to the header
level shown in the toc), and spent some effort setting up image
exporting so the sizing is acceptable.